字段调整，状态枚举内容调整

+++ Insurance/src/main/java/com/creditcloud/insurance/model/InsuranceProduct.java
@@ -110,7 +110,7 @@ public class InsuranceProduct extends BaseObject implements InsuranceConstant {
     /**
      * 预期收益率（平台提供）
      */
-    private int expectedRate;
+    private BigDecimal expectedRate;
     
     /**
      * 平台内状态
